Add localization-based global-position mimic with per-entry XY origins
Integrate estimator odometry for the K1 Glopodanamite policy and capture fresh robot/reference XY origins and yaw alignment on every mimic entry. Preserve continuous localization while walking between dances, ignore duplicate timestamps without extending freshness, and hand off to Velocity when localization is unavailable.

Match MJLab motion playback and action clipping, and add launch support, deployment documentation, and regression tests including ONNX reentry after translation and turning.

+++ ai_sapiens_sim2real/include/ai_sapiens_sim2real/policy/motion_playback.hpp
@@ -36,9 +36,9 @@ struct MotionPlayback
 
   // Absolute motion time to seek at this episode time, or nullopt once that
   // motion time passes time_end.
-  std::optional<float> seek_time(float episode_time) const
+  std::optional<double> seek_time(double episode_time) const
   {
-    const float motion_time = time_start + episode_time;
+    const double motion_time = time_start + episode_time;
     if (motion_time > time_end) {
       return std::nullopt;
     }
